% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_equality.R
\name{check_equality}
\alias{check_equality}
\title{Checks if two vectors of the same name of equal length have equal values}
\usage{
check_equality(column_name, datatable, keys)
}
\arguments{
\item{column_name}{name of column to be tested}

\item{datatable}{name of data table containing the columns to be tested}

\item{keys}{vector name of columns to be included as reference}
}
\value{
data table with only the tested columns and a mismatch column
}
\description{
This is a private function for dfcompare.
It requires a data table containing the name of the column and the
prefix "i." + name of the column to be compared with each other.
}
