% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_colnames.R
\name{common_colnames}
\alias{common_colnames}
\title{Searches for common column names between two data frames}
\usage{
common_colnames(source, target)
}
\arguments{
\item{source}{data frame of source}

\item{target}{data frame of target}
}
\value{
a vector of common column names
}
\description{
This is a private function for dfcompare.
}
