% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{df_compare}
\alias{df_compare}
\title{Compare two data frames}
\usage{
df_compare(source, target, keys)
}
\arguments{
\item{source}{data frame of the source}

\item{target}{data frame of the target}

\item{keys}{a string containing a key. Pass a character vector for multiple keys. The keys will be used in a data.table join using on =.}
}
\value{
a list of data.tables for each column and the corresponding mismatches
}
\description{
This function compares two data frames.  It checks if the source and target
variables are data frames, and if there are rows in both,
and if the keys are a character vector, and that they keys
exist in both source and target data frames.
}
