% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch_datatypes.R
\name{mismatch_datatypes}
\alias{mismatch_datatypes}
\title{Searches for mismatching datatypes between two data frames}
\usage{
mismatch_datatypes(source, target)
}
\arguments{
\item{source}{data frame of source}

\item{target}{data frame of target}
}
\value{
a data frame of column names with mismatching data types
}
\description{
This is a private function for dfcompare.
}
