% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch_colnames.R
\name{mismatch_colnames}
\alias{mismatch_colnames}
\title{Searches for uncommon column names between two data frames}
\usage{
mismatch_colnames(source, target)
}
\arguments{
\item{source}{data frame of source}

\item{target}{data frame of target}
}
\value{
a list of data frames containing uncommon column names
}
\description{
This is a private function for dfcompare.
}
