% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mismatch_datatypes.R
\name{mismatch_datatypes}
\alias{mismatch_datatypes}
\title{Searches for mismatching datatypes between two data frames}
\usage{
mismatch_datatypes(source, target, src_name = "source", tgt_name = "target")
}
\arguments{
\item{source}{data table of source}

\item{target}{data table of target}
}
\value{
a data table of column names with mismatching data types
}
\description{
This is a private function for dfcompare.
}
